mssmt: skip race tests to avoid long CI runs
